// frontend/TalentTribe/src/app/models/interview.model.ts
export interface Interview {
    interviewId: number;
    applicationId: number;
    interviewDate?: string;
    interviewType?: string;         // Optional
    interviewLink?: string;         // Optional
    interviewLocation?: string;     // Optional
    interviewFeedback?: string;     // Optional
    
  }
  